Add files via upload
For use with a WAMP server such as https://www.uniformserver.com/ if you don't already have a WAMP server set-up. Some code changes are required for use on a LAMP server set-up. You're a big boy... or girl. You can figure it out. Google is my friend. It can be yours too!